Add REST API project and integrate with MAUI app
- Add FastTrak.Api project with menu/restaurant endpoints
- Create RestaurantApiService for HTTP data fetching
- Standardize nutrition types to decimal across projects
- Add API hardening: CORS, error handling, logging
- Update documentation and project structure

+++ Models/MenuItem.cs
@@ -18,9 +18,9 @@ namespace FastTrak.Models
 
         public int Calories { get; set; }
 
-        public double Protein { get; set; }
-        public double Carbs { get; set; }
-        public double Fat { get; set; }
+        public decimal Protein { get; set; }
+        public decimal Carbs { get; set; }
+        public decimal Fat { get; set; }
         public int Sodium { get; set; }
 
         // Category (optional)
